cloud_functions/ai_query_api/requirements.txt
functions-framework==3.*
flask==2.*
requests==2.31.0
google-auth==2.*
google-auth-httplib2==0.1.0
google-cloud-aiplatform==1.38.1
google-cloud-storage==2.13.0
Pillow==10.2.0  # Required for image processing
